third_party/blink/web_tests/platform/generic/fast/events/flags-unset-on-init-event-expected.txt

An event's 'stop propagation', 'stop immediate propagation' and 'canceled' flags should be unset in initEvent

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS e.defaultPrevented is true
Running e.initEvent again.
PASS e.defaultPrevented is false
PASS successfullyParsed is true

TEST COMPLETE
